pyproject.toml
[project]
name = "consilium"
version = "0.1.0"
description = "Multi-Agent LLM Debate System - Inspired by Karpathy's llm-council"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-aws>=1.3.1",
    "langchain>=1.2.10",
    "python-dotenv>=1.2.2",
    "click>=8.1.0",
    "gradio>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]
